requirement.txt
mcp[cli]
requests
python-dotenv

